requirements.txt
Django~=3.2.7
django-redis~=5.0.0
djangorestframework~=3.12.4
python-decouple~=3.4
python-dotenv~=0.19.0
